Refactor further
