+++ packages/common/src/utils/index.ts
@@ -25,4 +25,4 @@ export { License } from './creativeCommons'
 export * from './array'
 export * from './chatUtils'
 export * from './twitter'
-export * from './premiumContent'
+export * from './streaming'
