+++ packages/mobile/src/components/audio/Audio.tsx
@@ -1,6 +1,6 @@
 import { useRef, useEffect, useCallback, useState } from 'react'
 
-import { generateUserSignature, ID, Nullable, QueryParams, Track } from '@audius/common'
+import { getQueryParams, ID, Nullable, QueryParams, Track } from '@audius/common'
 import {
   removeNullable,
   playbackRateValueMap,
@@ -282,18 +282,9 @@ export const Audio = () => {
         if (gatedQueryParamsMap[trackId]) {
           queryParamsMap[trackId] = gatedQueryParamsMap[trackId]
         } else if (isGatedContentEnabled) {
-          const { data, signature } = await generateUserSignature(audiusBackendInstance)
           const premiumContentSignature =
             premium_content_signature || premiumTrackSignatureMap[trackId]
-          queryParamsMap[trackId] = {
-            user_data: data,
-            user_signature: signature
-          }
-          if (premiumContentSignature) {
-            queryParamsMap[trackId].premium_content_signature = JSON.stringify(
-              premiumContentSignature
-            )
-          }
+          queryParamsMap[trackId] = await getQueryParams({ audiusBackendInstance, premiumContentSignature })
         }
       }
